requirements.txt
flask==2.3.0
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1
pyyaml==6.0
